Refactor handlers for link method
